Refactor to use game state object
